fix: refactor components structure
