+++ web/src/components/elements/panel/PanelAction/PanelAction.css
