+++ web/src/components/elements/panel/PanelAction/index.ts
@@ -1,0 +1,1 @@
+export * from './PanelAction';
