+++ web/src/components/elements/panel/PanelHeader/index.ts
@@ -1,0 +1,1 @@
+export * from './PanelHeader';
