--- web/src/components/elements/tabs/TabHeader/TabHeader.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-interface Props {}
-
-export const TabHeader: React.FC<Props> = () => {
-  return (
-    <></>
-  )
-}
