--- web/src/components/elements/tabs/TabLabel/TabLabel.tsx
@@ -1,1 +1,0 @@
-export {}
